+++ datatypes.h
@@ -1208,6 +1208,7 @@ public:
     Q_PROPERTY(QString description_md MEMBER description_md)
     Q_PROPERTY(QByteArray lispData MEMBER lispData)
     Q_PROPERTY(QString qmlFile MEMBER qmlFile)
+    Q_PROPERTY(QString pkgDescQml MEMBER pkgDescQml)
     Q_PROPERTY(bool qmlIsFullscreen MEMBER qmlIsFullscreen)
     Q_PROPERTY(bool isLibrary MEMBER isLibrary)
     Q_PROPERTY(bool loadOk MEMBER loadOk)
@@ -1226,6 +1227,7 @@ public:
     QString description_md;
     QByteArray lispData;
     QString qmlFile;
+    QString pkgDescQml;
     bool qmlIsFullscreen;
     bool isLibrary;
     bool loadOk;